Refactor and remove unnecessary files
Renamed multiple method names for increased clarity and conciseness.
Changed the usage of Lists in some instances to IReadOnlyList to
emphasize those instances as read-only. Combined some interfaces that
provided a similiar function, which may need to be further modified
later depending on the structure of the program. Decoupled some classes
to prepare for application layering later on.