Add repositories to load game data
Repositories use Json.Net to parse json files for game data objects.
Most game entities were modified to provide more ease of access to
those looking to add more objects to the game. The data files included
are just test data and must be included in the debug folder to correctly
load into the game.

There is now a distinction between immutable and mutable game data.
A Character is an example of a mutable instance of game data, the
CharacterBase, the template that character is based off of, is
immutable.