Implement attacks, spells, and status effects
Attacks, spells and status effects have been implemented along with
their delayed counterparts. As a consequence, consumable items, which at
their core are limited-usage spells, are also implemented. The
CharacterController class may be redundant. It's original purpose was
to handle any updates to a Character's properties and prevent any
unauthorized changes.

Concerning the UI, some display bugs have been fixed. The game now
successfully implements the usage of different "center of target"
positions for actions, which is the position where the player uses to
aim an action. Without this, it would have been impossible to have
some target combinations.