Add partial unit test for UI assembly
Added unit testing for the action panel and combat formation UIs. In
the process, found and fixed some minor bugs. Unit testing might not be
a great idea for highly malleable code such as the UI which might
undergo several breaking changes which might each require changes to
the unit testing code as well.