Homogenize attacks, spells, and skills
Before now, attacks, spells, and skills shared many similiarities in
their implementations and designs, but had several different methods
and properties to perform the same tasks. This violated the DRY rule
and gave the duty of differentiating between those actions to the
programmer. Now attacks, spells, and skills have most of their unique
properties moved to the shared base class ActionBase. What results is
that now attacks, spells, and skills share one implementation through
the base class and the duty of determining the differences between the
actions belongs to the game designer instead of the programmer.