Refactor damage calculation into seperate class
The damage calculation performed in the ActionController and
StatusController classes are needed in many classes outside of those
controllers, especially in the case of AI, where the AI will be using
those damage calculations to make decisions. The DamageCalculator
class is static because it should not be storing any data and should
only be used to get damage and healing amounts.