Add threat generation to actions and statuses
Threat is a number that represents how threatening a character is to
the AI. A player character with high threat is much more likely to be
targeted by the AI than a player character with low threat. Threat
is generated from actions and status effects. The AI characters can also
gain threat but will not be used to determine actions against other AI
characters at this point in time.

Originally, threat was going to be directly based on the amount of
damage or healing a character did. The issue that arises is that
actions and status effects that cause no damage but can still
significantly sway the course of the battle will cause much less or
much more threat than actions and status effects that deal damage. This
is because actions and status effects have a base threat amount
generated on every use, which could be high or low depending on how
much health the average character in the game has. The solution is to
calculate threat based on how much an action or status effect has
changed a character's health in percentage. That way threat numbers can
be smaller and more manageable.