Fix critical errors with AI
Firstly, threat generation was negative whenever an attack was
being made by the player. So an attack would reduce threat instead of
increase it, while a heal would increase threat. The temporary solution
is to use absolute values so that heals and attacks would both increase
threat. The drawback is that heals to enemy characters will still
generate threat instead of reducing it.

Secondly, the AI would still target dead player characters for it's
attacks. The fix was to remove dead player characters from the AI's
playercharacter list.

Lastly, the AI didn't handle the case of there being only 2 player
characters when it uses the median to calculate which player character
is more threatening. The fix was simply to bypass the using the median
and give a higher priority value to the more threatening of the two.