Fix and add test for static-target actions
Actions that cannot change target positions were not handled correctly
before. Previously, the AI checked all positions on the battlefield for
the highest priority position even if the action could not change target
positions, which is a waste of computations. Also, the AI translated
positions were not applied to actions that can't change target positions
so if somehow a different target position was chosen for an action,
other than the default, that would change the targets for the action.