Add script to include data files in build
Added a post-build event command line script to copy data files from
the solution directory to the build directory. A problem exists in that
in order to see the changes made to the data files, you must rebuild
the project within Visual Studio to have it recopy over the changes.