Implement gradual health changes for ui
Implemented gradual health changes for the UI. In order to implement it,
several changes had to be made. The big problem with the previous system
was that the UI was very tightly binded to the data from the controller,
so changing health from the controller updated the health values in the
UI instantaneously at the end of the turn. Having the controller
gradually change health was not an option, so that mean the UI layer
needed to have it's own version of the characters in combat. With the UI
layer owning it's own version of characters, it could apply change in
health updates gradually without affecting the actual health of the
characters.

A few events were added to increase the interactivity between the
controller and UI layers, but a lot more should be added to detail
ingame events that must be visualized for players.

UI files moved to a combat folder within the UI project for more
structure.