Finetune gradual health updates
A pause is added at the end of every health update to give the player
a chance to understand the changes happening.

Previously, the ActionController would calculate and do damage to a
dead character, which only surfaced as a bug when testing area of effect
damage against living and dead targets. The gradual health updates
would pick up the damage even though a character was dead before and
after the action's effects. It changed the UI version of the character
even though it should not be changed. The fix was to not add dead
characters that remain dead to the health changed event arguments. This
will allow resurrection to be implemented later on.