Add defaults to combatui properties
Almost all of the properties in the CombatUI reference a dictionary
that stores PlayerCharacterDefaults. The properties only return values
corresponding to the currently active character. Problems may then occur
whenever some code tries to use the properties with a key that does not
exist in the dictionary, such as in the case of an AI character. The fix
is to provide key checking, and in the case that a key does not exist
in the dictionary, provide a suitable default return value.