Add frameupdates to AI turns
Added multiple frame updates to the UI on AI turns to help the player
visualize which targets the AI has chosen. Also changed the UI
component rendering to allow for the rendering of AI targets during the
AI's turn.

Some changes were made to fix the UI bug where a player character's
actions were shown even though it was the enemy turn.