Implement way to skip UI updates
Some players find it very convenient to skip the long process of UI
updates, such as skipping the gradual change of health. In order to
implement update skipping, the UI updates had to occur in a different
thread to be able to also listen for user input. On the correct input,
the UI will skip its gradual updates and instantly finish.

One concern is that the variable _skipUpdating is not atomic and is
being set and read in seperate threads. However, since the UI updates so
quickly and rendering an extra frame is not a problem, it is not worth
implementing a lock which would dramatically slow the UI rendering.