Refactor a massive amount of the codebase
Prior to this commit, two classes started to show signs of being god
objects: the CombatUI and CombatController classes. Both were becoming
too bloated as more responsiblities piled onto the classes. The
CombatController was the only class the UI was able to use to receive
all the data from the backend which is bad design in the long term
scheme of things. The CombatUI just held way too much information, from
knowing how to construct UI components to handling user input. Therefore
both of these classes were split into smaller classes which can each
handle a single responsiblity.

Very many classes and variables were renamed to more concisely convey
what data they store. This fixes the problem the codebase used to have
where similiar terms were used to describe unsimiliar entities. For
example, one problem was the difficulty in differentiating between a
SubAction and an Action. It was confusing to work with both entities
with such similiar and undiscerning names. SubActions are now renamed
to Actions, and Actions to Commands to more effectively reflect their
purpose.

GetModifiedTargetPositions was removed from the CombatUI and placed into
the Shared project. It has also been simplified to be human-readable.
There were also very many redundant calls to the function in the
CombatUI that were removed.

A few bugs were fixed such as DelayedStatuses not supplying the
ThreatController with a source character, and the input not properly
flushing when a UI update is skipped.