Implement caching for UI components
UI components do a lot of string concatenation and can be very costly on
performance, especially in case of multiple frame updates. Most of the
time those UI components are rerendering the same data over and over and
is very wasteful in terms of CPU resources. Now each UI component caches
their most recently rendered data and reuses a render if the next set of
render data is identical to the cached render data. There are still
improvements that can be made to this elementary caching system such as
saving multiple renders in components that have few state changes.

InformationPanelUI is now considered obsolete and will be removed in a
future version.