Implement speed change event
The StatusController will now invoke a CharacterSpeedChanged event
whenever a character has a status effect applied to it that changes it's
speed. The CombatStateHandler will now make any changes to the round
order if a speed-changed character surpasses or falls below the speed of
another character.

Fixed a few bugs where the CharactersHealthChanged event was being
invoked even though no characters had their health changed.

The Thread.Sleep call in CombatUI was moved to the EndOfTurn evenhandler
from the CharactersHealthChanged eventhandler to allow every action,
regardless if whether or not if modified health, to have a pause before
starting the next turn.