Prevent slowed characters acting after waiters
Waiting characters knowingly put themselves at the end of the round
order after all other non-waiting characters. The problem is that
whenever a character is slowed, it would be moved to the index in the
round order where it belongs speed wise; which would be after the
waiting characters. Game logic dictates that no matter how slow a
character is or becomes, a waiting character still acts after them. This
is now implemented by having the CombatStateHandler keep track of where
in the round order the waiting characters exist, and preventing slowed
characters from moving past those positions when the round order is
changed due to speed changes.

Made a minor fix to prevent the last character in the round from waiting
indefinitely.