Move target verification to combat controller
Previously the UI passed into the CombatController the target positions
an action a player selected would affect. This was prone to abuse as the
target positions were never checked for incorrect positions. Now,
instead of passing in the target positions, the UI will pass in a single
target position: the center target of the action being used. The
CombatController will then use that to construct the correct target
positions. This will solve problems in the long run if there are ever
frontend and backend inconsistencies in terms of targeting.