Implement basic combat log panel
A combat log panel wasn't going to fit in nicely to any of the remaining
space in the current console window size, so the width was extended by
40 units. Now there's extra space on the right side of the console
window that should be filled up with extra components. The current size
of the combat log panel is 40x16, which errs on the smaller side. It's
current position on the bottom right side of the console is fine but it
doesn't feel as natural to start at when most other games implement
their combat logs on the bottom left.

Currently, the combat log panel receives combat messages through events
that the ActionController and StatusController invoke whenever an action
occurs in combat. A class called CombatLoggableEventArgs was created
for other events to inherit in the case that their events should be
logged to the combat log panel.