Fix bugs with action and status controllers
The ActionController invoked the CharactersHealthChanged event and
characters that were dead before the action was started were included in
the list of characters that had their health changed. Simply discluding
any dead characters would prevent any characters that were resurrected
from showing on the combat log. So a check is now being made to see if
an action resurrects the dead before including the dead in the list of
health changed characters.

The status controller would invoke the StatusEffectApplied event even
when all the target characters were dead, but with an empty log message.
Now, the ActionController decides if it's necessary to apply any status
effects if all the characters died as a result of the action.

Lastly, whenever a delayed action with a status effect was used, the
status effect would get applied before the damage and healing from the
delayed action. This makes the combat log display the status effect from
the delayed action being displayed before the action is executed.Game
logic dictates it be the other way around. Now, the flow of combat is
slightly altered; A character would start it's turn, first applying any
healing or damage from status effects from previous turns, then removing
expired status effects, then start any delayed actions the character has
stored, and finally apply delayed statuses as a result of that action.
With that change, the message log now properly displays the delayed
action being executed before the status effect.