Implement character panel basics
Implemented the basics of the character panel, which expands upon the
previous CharacterDetailsPanel. Now that the CharacterPanel has taken
over the functions of the CharacterDetailsPanel, the
CharacterDetailsPanel is now marked as obsolete. This means that the
CategoryDetailsPanel and ActionDetailsPanel no longer share a spot on
the UI with the CharacterDetailsPanel. When previously the
CharacterDetailsPanel would be active in the FormationPanel, the
ActionDetailsPanel will instead stay active in it's place.
The CharacterPanel, which replaces the CharacterDetailsPanel
takes up the empty space on the right side of the console window.

The CharacterPanel has the ability to display multiple targets, which
should be useful to show the player the different targets being
selected. A feature to add in later is to be able to press tab to toggle
between the different targets to see more details about the characters.

In order to implement the CharacterPanel, the FormationPanel had to be
fixed to allow for size scaling. The CharacterPanel now resizes itself
to be the same height as the FormationPanel.