Create subpanels to handle seperate stats
Created subpanels that render different stats of a character. They
implement the IReceiveInputPanel interface because later they will be
used with the status command to allow the player to select a stat that
they want to view in more detail.

All stats have been successfully implemented into the CharacterPanel
except for status effects.