Refactor UI layer
Reworked the flow of logic for the UI layer. Before, most of the input
handling was done by the UserInput class. The flaw in this approach was
that code was too tightly coupled from the UserInput class to the
UIContainer class to all of the UI panels. It meant both the
UserInput and UIContainer classes had to know how each UI component is
constructed to handle their rendering.

Now, each UI panel knows how to render itself and has the means, through
dependency injection, to supply itself with data. Some UI panels also
have the means to handle user input, which will come in handy when
implementing new features like the status command.

Implemented custom scopings with ninject. Many of the classes are used
like a service and while they shouldn't be made singletons, there should
only be one instance of them per combat instance. Since the CombatUI is
the root of the combat instance, a new scope should be created per
CombatUI instance to make use of the custom scoping.