Fix bugs related to formation panel
Two bugs were introduced when removing the tight coupling between the
UIContainer and the FormationPanel. The first bug crashed the program
when it was the AI's turn, and the second bug crashed the program after
using an item.

The fix for the first bug was to allow the UIStateTracker to keep track
of whether it is currently the player's turn or not. This allowed the
UI panels, namely the ActionDetailsPanel, which depended on the active
character, to only request information during the player's turn. This
stopped the crashes because the UIStateTracker changes data based on
which player character's turn it currently is. This removes null objects
because the UIStateTracker doesn't store data for ai character turns.

The second bug was fixed by switching the orders in which the
Consumable item was removed from the inventory of the consumer. This
prevents a NullReferenceException in cases where the item is removed
before it could be displayed in the UI.