Implement subpanel key press handling
Subpanels within the CharacterPanel are now implemented to handle input
from the player. In order to avoid having the UIStateTracker keep track
of every single subpanel the player is in, the CharacterPanel will be
responsible for handling which of its subpanels is active. This required
the CharacterPanel to handle the enter and escape key presses, but also
caused many problems where the enter/escape key presses were being
handled in multiple places, causing multiple interactions.

The solution to this problem and a logical design choice was to limit
each key press to being handled once. This means that user input would
theoretically have a chance to be handled by the lowest level UI
subpanels before they can be handled by containing panels or the
UIContainer. If one component handles the user input, all other
components would ignore that input.