Add StatsDetailsPanel that displays stat info
Added a StatsDetailsPanel that displays extra information for each
character stat to the player. The stat descriptions are constructed from
the statdescriptions.json file and are retrieved by id by the
StatDescriptionController.

In order to relay data from the CharacterPanel to the StatsDetailsPanel,
several events have been added to notify listeners about focus changes
and activeness changes in the CharacterPanel and its subpanels.