Implement ScreenBuffer to reduce flickering
The ScreenBuffer was successfully implemented. The differences between
the previous method of printing to the Console and the method the
ScreenBuffer uses is that the ScreenBuffer stores the entire next frame
to print and the ScreenBuffer does not call Console.Clear anymore.
Instead, the ScreenBuffer writes directly over the old frame,
eliminating the flickering that occurs as a result of clearing the
Console.